Expand Travis matrix w/ minimum duplication
Before the change I made in
https://github.com/alanjds/drf-nested-routers/commit/eb4f443230abfefd2c3a065abe1d7dc3cc02cb25,
we had Travis doing a lot of parallel tests (a good thing) but the
matrix had to be defined in both Travis and Tox (a bad thing).  After
that commit, the matrix was only in Tox (a good thing) but it cut down
on the parallel testing in Travis (a bad thing).

This commit is an attempt to balance the two things.  Tox is still the
holder of the overall matrix but we duplicate a minimal amount of
information to travis to get more jobs running at once.